Added less compiler, autoprefixer, and removed obsolete code.
