--- consts.js
@@ -1,7 +1,0 @@
-const defaultOptions = {
-  generateScopedName: '[local]',
-};
-
-module.exports = {
-  defaultOptions,
-};
