+++ examples/Button.h.less
@@ -1,0 +1,31 @@
+.oc-button__overlay() {
+  background-color: rgba(255, 255, 255, 1);
+  content: '';
+  height: 100%;
+  left: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+  z-index: 2;
+}
+
+.oc-button__overlay--light() {
+  .oc-button__overlay();
+  background-color: rgba(255, 255, 255, 0.36);
+  z-index: 1;
+}
+
+.oc-button__overlay--lighten() {
+  .oc-button__overlay();
+  background-color: rgba(255, 255, 255, 0.64);
+}
+
+.oc-button__overlay--dark() {
+  .oc-button__overlay();
+  background-color: rgba(0, 0, 0, 0.08);
+}
+
+.oc-button__paper-shadow() {
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.38);
+  transition: box-shadow 80ms ease-in-out;
+}
