+++ examples/Button.less
@@ -1,0 +1,115 @@
+@import 'Button.h.less';
+
+.oc-button__alt-content {
+  align-items: center;
+  display: flex;
+  height: 100%;
+  justify-content: center;
+  left: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+}
+
+.oc-button {
+  align-items: center;
+  border: none;
+  border-radius: 2px;
+  background-color: #fff;
+  display: inline-flex;
+  font-size: 14px;
+  font-weight: 600;
+  padding: 4px 16px;
+  position: relative;
+  overflow: hidden;
+}
+
+.oc-button--content-before {
+  padding-left: 8px;
+}
+
+.oc-button--content-after {
+  padding-right: 8px;
+}
+
+.oc-button--content-empty {
+  padding-left: 4px;
+  padding-right: 4px;
+}
+
+.oc-button:focus {
+  outline: none;
+}
+
+.oc-button:hover {
+  cursor: pointer;
+}
+
+.oc-button:hover::before {
+  .oc-button__overlay--light();
+}
+
+.oc-button:hover:active::before {
+  content: initial;
+}
+
+.oc-button:focus:not(.disabled)::after {
+  .oc-button__overlay--light();
+}
+
+.oc-button:focus::before {
+  .oc-button__overlay();
+  background-color: transparent;
+  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.32);
+}
+
+.oc-button--active::before {
+  .oc-button__overlay();
+  background-color: transparent;
+  box-shadow: none;
+}
+
+.oc-button--active::after,
+.oc-button--active:focus:not(.oc-button--disabled)::after,
+.oc-button:focus:active::after {
+  .oc-button__overlay--dark();
+}
+
+.oc-button__content {
+  display: inline-flex;
+  align-items: center;
+}
+
+.oc-button__children {
+  align-items: center;
+  display: inline-flex;
+}
+
+.oc-button__delimiter {
+  width: 0.4em;
+}
+
+.oc-button--disabled:hover::before,
+.oc-button--disabled:focus::before,
+.oc-button--disabled:focus:active::before,
+.oc-button--disabled::before {
+  content: initial;
+}
+
+.oc-button--disabled::after,
+.oc-button--disabled:focus:active::after,
+.oc-button--disabled:hover:focus:active::after {
+  .oc-button__overlay--lighten();
+}
+
+.oc-button__disabled:hover {
+  cursor: not-allowed;
+}
+
+.oc-button__label::selection {
+  background-color: transparent;
+}
+
+.oc-button--paper {
+  .oc-button__paper-shadow();
+}
