+++ examples/example1.js
@@ -1,3 +1,5 @@
+import './Button.less';
+
 import './example-less.less'
 
 import './example1.css'
