+++ package.json
@@ -1,17 +1,17 @@
 {
   "name": "babel-plugin-transform-import-styles",
-  "version": "0.0.2",
-  "description": "Injects imported styles (css-modules) into js",
-  "main": "index.js",
+  "version": "0.0.3",
+  "description": "Injects imported styles (.css or .less) into js",
+  "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build-examples": "babel examples/ -d examples-dist/ --plugins ../index.js",
-    "postversion": "git push && git push --tags && npm publish"
+    "build-examples": "babel examples/ -d examples-dist/ --plugins ../src/index.js",
+    "publish": "npm publish"
   },
   "keywords": [
     "babel",
     "css",
-    "css-modules"
+    "less"
   ],
   "repository": {
     "type": "git",
@@ -22,17 +22,13 @@
     "access": "public",
     "registry": "https://registry.npmjs.org"
   },
-  "author": "Alex Mextner <invntrm@gmail.com>",
+  "author": "Egor Stambakio <stambakio@scand.com>",
   "license": "MIT",
   "dependencies": {
+    "autoprefixer": "8.4.0",
     "less": "3.0.2",
     "load-styles": "2.0.0",
-    "postcss": "6.0.19",
-    "postcss-icss-selectors": "2.0.3",
-    "postcss-modules-extract-imports": "1.1.0",
-    "postcss-modules-resolve-imports": "1.3.0",
-    "postcss-modules-scope": "1.1.0",
-    "postcss-modules-values": "1.3.0",
+    "postcss": "6.0.22",
     "require-resolve": "0.0.2"
   },
   "peerDependencies": {
