+++ src/css-import-visitor.js
@@ -1,10 +1,17 @@
 const { readFileSync } = require('fs');
 const path = require('path');
 const requireResolve = require('require-resolve');
-const less = require('less');
 const postcss = require('postcss');
 
-const { defaultOptions } = require('./consts');
+function errorBoundary(cssFile, cb) {
+  try {
+    cb();
+  } catch (e) {
+    // debugger; // eslint-disable-line no-debugger
+    console.error(new Error(`babel-plugin-transform-import-styles: ${ cssFile }: ${ e.message }`));
+    throw e;
+  }
+}
 
 /**
  * Visitor for `import '*.css'` babel AST-nodes
@@ -16,25 +23,27 @@ function CssImport(cb) {
       if (!node.source.value.endsWith('.css') && !node.source.value.endsWith('.less')) return;
 
       const { src } = requireResolve(node.source.value, path.resolve(file.opts.filename));
-      const content = readFileSync(src, 'utf8');
 
-      let css = content;
+      console.log({ src })
+
+      let css;
 
       if (node.source.value.endsWith('.less')) {
         // unfortunately babel is completely sync
         // we need to block while we compile .less
         css = require('child_process').execSync(
-          `node compile-less.js`,
+          `node ${path.join(__dirname, 'compile-less.js')} ${src}`,
           {
             cwd: __dirname,
-            input: content,
             encoding: 'utf8'
           }
         )
+      } else {
+        css = readFileSync(src, 'utf8');
       }
 
       // TODO: load postcss options and plugins
-      const options = { ...defaultOptions, ...opts };
+      const options = { generateScopedName: '[local]', ...opts };
 
       cb({
         babelData,
@@ -49,12 +58,3 @@ function CssImport(cb) {
 
 module.exports = CssImport;
 
-function errorBoundary(cssFile, cb) {
-  try {
-    cb();
-  } catch (e) {
-    debugger; // eslint-disable-line no-debugger
-    console.error(new Error(`babel-plugin-transform-import-styles: ${ cssFile }: ${ e.message }`));
-    throw e;
-  }
-}
